@@ -1,60 +1,42 @@
-module github.com/walteh/retab
+module github.com/walteh/retab/v2
 
-go 1.22.4
+go 1.23.4
 
 require (
 	github.com/apparentlymart/go-textseg/v13 v13.0.0
-	github.com/bufbuild/protocompile v0.13.0
+	github.com/bmatcuk/doublestar/v4 v4.7.1
+	github.com/bufbuild/protocompile v0.14.1
 	github.com/editorconfig/editorconfig-core-go/v2 v2.6.2
-	github.com/go-faster/errors v0.7.1
 	github.com/hashicorp/go-multierror v1.1.1
-	github.com/hashicorp/hcl/v2 v2.20.1
-	github.com/k0kubun/pp/v3 v3.2.0
-	github.com/kylelemons/godebug v1.1.0
-	github.com/rs/xid v1.5.0
-	github.com/rs/zerolog v1.32.0
-	github.com/santhosh-tekuri/jsonschema/v5 v5.3.1
+	github.com/hashicorp/hcl/v2 v2.23.0
+	github.com/rs/zerolog v1.33.0
 	github.com/spf13/afero v1.11.0
-	github.com/spf13/cobra v1.8.0
-	github.com/stretchr/testify v1.9.0
-	github.com/walteh/snake v0.48.0
-	github.com/walteh/terrors v0.19.0
-	github.com/walteh/yaml v0.0.0-20240212175823-ac6e39d6b02d
-	github.com/zclconf/go-cty v1.14.4
-	gitlab.com/tozd/go/errors v0.8.1
+	github.com/spf13/cobra v1.8.1
+	github.com/stretchr/testify v1.10.0
+	gitlab.com/tozd/go/errors v0.10.0
 	go.uber.org/multierr v1.11.0
-	golang.org/x/mod v0.17.0
 )
 
 require (
-	github.com/agext/levenshtein v1.2.3 // indirect
+	github.com/agext/levenshtein v1.2.1 // indirect
 	github.com/apparentlymart/go-textseg/v15 v15.0.0 // indirect
-	github.com/chzyer/readline v1.5.1 // indirect
-	github.com/davecgh/go-spew v1.1.2-0.20180830191138-d8f796af33cc // indirect
-	github.com/fatih/color v1.16.0 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/google/go-cmp v0.6.0 // indirect
-	github.com/hashicorp/errwrap v1.1.0 // indirect
+	github.com/hashicorp/errwrap v1.0.0 // indirect
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
-	github.com/k0kubun/colorstring v0.0.0-20150214042306-9440f1994b88 // indirect
-	github.com/kr/pretty v0.3.1 // indirect
-	github.com/manifoldco/promptui v0.9.0 // indirect
 	github.com/mattn/go-colorable v0.1.13 // indirect
 	github.com/mattn/go-isatty v0.0.20 // indirect
-	github.com/mattn/go-runewidth v0.0.15 // indirect
-	github.com/mitchellh/go-wordwrap v1.0.1 // indirect
-	github.com/olekukonko/tablewriter v0.0.5 // indirect
+	github.com/mitchellh/go-wordwrap v0.0.0-20150314170334-ad45545899c7 // indirect
 	github.com/pkg/errors v0.9.1 // indirect
-	github.com/pmezard/go-difflib v1.0.1-0.20181226105442-5d4384ee4fb2 // indirect
-	github.com/rivo/uniseg v0.4.7 // indirect
-	github.com/rogpeppe/go-internal v1.12.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
 	github.com/stretchr/objx v0.5.2 // indirect
-	golang.org/x/sync v0.7.0 // indirect
-	golang.org/x/sys v0.20.0 // indirect
-	golang.org/x/text v0.15.0 // indirect
-	golang.org/x/tools v0.21.0 // indirect
-	google.golang.org/protobuf v1.34.1 // indirect
-	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
+	github.com/zclconf/go-cty v1.13.0 // indirect
+	golang.org/x/mod v0.16.0 // indirect
+	golang.org/x/sys v0.15.0 // indirect
+	golang.org/x/text v0.14.0 // indirect
+	golang.org/x/tools v0.13.0 // indirect
+	google.golang.org/protobuf v1.34.2 // indirect
 	gopkg.in/ini.v1 v1.67.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
