@@ -1,4 +1,4 @@
-// Code generated by mockery v2.42.0. DO NOT EDIT.
+// Code generated by mockery v2.50.1. DO NOT EDIT.
 
 package mockery
 
@@ -17,7 +17,7 @@ func (_m *MockConfiguration_format) EXPECT() *MockConfiguration_format_Expecter 
 	return &MockConfiguration_format_Expecter{mock: &_m.Mock}
 }
 
-// IndentSize provides a mock function with given fields:
+// IndentSize provides a mock function with no fields
 func (_m *MockConfiguration_format) IndentSize() int {
 	ret := _m.Called()
 
@@ -62,7 +62,7 @@ func (_c *MockConfiguration_format_IndentSize_Call) RunAndReturn(run func() int)
 	return _c
 }
 
-// OneBracketPerLine provides a mock function with given fields:
+// OneBracketPerLine provides a mock function with no fields
 func (_m *MockConfiguration_format) OneBracketPerLine() bool {
 	ret := _m.Called()
 
@@ -107,7 +107,7 @@ func (_c *MockConfiguration_format_OneBracketPerLine_Call) RunAndReturn(run func
 	return _c
 }
 
-// TrimMultipleEmptyLines provides a mock function with given fields:
+// TrimMultipleEmptyLines provides a mock function with no fields
 func (_m *MockConfiguration_format) TrimMultipleEmptyLines() bool {
 	ret := _m.Called()
 
@@ -152,7 +152,7 @@ func (_c *MockConfiguration_format_TrimMultipleEmptyLines_Call) RunAndReturn(run
 	return _c
 }
 
-// UseTabs provides a mock function with given fields:
+// UseTabs provides a mock function with no fields
 func (_m *MockConfiguration_format) UseTabs() bool {
 	ret := _m.Called()
 
